+++ src/RequestLocation/HeaderLocation.php
@@ -64,15 +64,37 @@ class HeaderLocation extends AbstractLocation
      */
     private static function prepareHeaderValue($value)
     {
+        if (is_string($value)) {
+            return $value;
+        }
+
         if (is_scalar($value)) {
+            trigger_deprecation(
+                'guzzlehttp/guzzle-services',
+                '1.6',
+                'Passing %s as a header location value is deprecated; guzzlehttp/guzzle-services 2.0 requires string|string[].',
+                get_debug_type($value)
+            );
+
             return (string) $value;
         }
 
         if (is_array($value)) {
             foreach ($value as $key => $item) {
+                if (is_string($item)) {
+                    continue;
+                }
+
                 if (!is_scalar($item)) {
                     throw new \InvalidArgumentException('Header location values must be scalar or an array of scalars.');
                 }
+
+                trigger_deprecation(
+                    'guzzlehttp/guzzle-services',
+                    '1.6',
+                    'Passing %s inside a header location value array is deprecated; guzzlehttp/guzzle-services 2.0 requires string|string[].',
+                    get_debug_type($item)
+                );
 
                 $value[$key] = (string) $item;
             }