+++ tests/RequestLocation/HeaderLocationTest.php
@@ -34,46 +34,16 @@ class HeaderLocationTest extends TestCase
     /**
      * @group RequestLocation
      */
-    public function testVisitsLocationSerializesScalarHeaderValues()
-    {
-        $location = new HeaderLocation('header');
-        $param = new Parameter(['name' => 'foo']);
-
-        $request = $location->visit(
-            new Command('foo', ['foo' => 123]),
-            new Request('POST', 'http://httbin.org'),
-            $param
-        );
-        $this->assertEquals([0 => '123'], $request->getHeader('foo'));
-
-        $request = $location->visit(
-            new Command('foo', ['foo' => true]),
-            new Request('POST', 'http://httbin.org'),
-            $param
-        );
-        $this->assertEquals([0 => '1'], $request->getHeader('foo'));
-
-        $request = $location->visit(
-            new Command('foo', ['foo' => false]),
-            new Request('POST', 'http://httbin.org'),
-            $param
-        );
-        $this->assertEquals([0 => ''], $request->getHeader('foo'));
-    }
-
-    /**
-     * @group RequestLocation
-     */
     public function testVisitsLocationSerializesArrayHeaderValues()
     {
         $location = new HeaderLocation('header');
-        $command = new Command('foo', ['foo' => ['bar', 123, true, false]]);
+        $command = new Command('foo', ['foo' => ['bar', 'baz']]);
         $request = new Request('POST', 'http://httbin.org');
         $param = new Parameter(['name' => 'foo']);
 
         $request = $location->visit($command, $request, $param);
 
-        $this->assertEquals([0 => 'bar', 1 => '123', 2 => '1', 3 => ''], $request->getHeader('foo'));
+        $this->assertEquals([0 => 'bar', 1 => 'baz'], $request->getHeader('foo'));
     }
 
     /**
@@ -127,27 +97,6 @@ class HeaderLocationTest extends TestCase
         $header = $request->getHeader('add');
         $this->assertIsArray($header);
         $this->assertEquals([0 => 'props'], $header);
-    }
-
-    /**
-     * @group RequestLocation
-     */
-    public function testAdditionalPropertiesSerializeScalarHeaderValues()
-    {
-        $location = new HeaderLocation('header');
-        $command = new Command('foo', ['foo' => 'bar']);
-        $command['add'] = 123;
-        $operation = new Operation([
-            'additionalParameters' => [
-                'location' => 'header',
-            ],
-        ]);
-        $request = new Request('POST', 'http://httbin.org');
-        $request = $location->after($command, $request, $operation);
-
-        $header = $request->getHeader('add');
-        $this->assertIsArray($header);
-        $this->assertEquals([0 => '123'], $header);
     }
 
     /**